requirements.txt
pytrends==4.9.2
google-generativeai==0.3.2
pandas==2.2.1
python-dotenv==1.0.1
numpy==1.26.4

